completed flag to remove confetti
